Added test files for Junit for password validation
These must be included in zip file, just putting them here as required.

Test case covered: Password length validation. This function will ensure that the password length is valid and the testing of this feature shows that the boolean value returned is accurate. Pictures of both cases are included in push.